bug fix on local_source

+++ cassandra_snapshotter/snapshotting.py
@@ -94,9 +94,9 @@ class Snapshot(object):
 
 class RestoreWorker(object):
     def __init__(self, aws_access_key_id, aws_secret_access_key, snapshot, local_source='',
-                 merge_dir='.', overwrite_local_source=False):
+                 merge_dir='.', overwrite_local_merge=False):
 
-        if not local_source or overwrite_local_source:
+        if not local_source or overwrite_local_merge:
             self.aws_secret_access_key = aws_secret_access_key
             self.aws_access_key_id = aws_access_key_id
             self.s3connection = S3Connection(aws_access_key_id=self.aws_access_key_id,
@@ -107,15 +107,15 @@ class RestoreWorker(object):
 
         self.local_source = local_source
         self.merge_dir = merge_dir
-        self.overwrite_local_source = overwrite_local_source
+        self.overwrite_local_merge = overwrite_local_merge
 
         if local_source and not os.path.exists(local_source) :
             os.makedirs(local_source)
 
         self.path_separator = os.path.sep
-        
-    def restore(self, keyspace, table, hosts, target_hosts):
 
+
+    def restore(self, keyspace, table, hosts, target_hosts):
         self._restore(keyspace, table, hosts, target_hosts)
 
 
@@ -181,7 +181,7 @@ class RestoreWorker(object):
         matcher_string = "(%(hosts)s).*/(%(keyspace)s)/(%(table)s)/" % dict(hosts='|'.join(hosts), keyspace=keyspace, table=table)
         self.keyspace_table_matcher = re.compile(matcher_string)
 
-        if self.local_source and not self.overwrite_local_source :
+        if self.local_source :
             logging.info("Restoring keyspace=%(keyspace)s, table=%(table)s, "
                      "from existing local data: %(local_dir)s " % dict(keyspace=keyspace,
                                                                            table=table, local_dir=self.local_source))
@@ -214,8 +214,8 @@ class RestoreWorker(object):
 
         logging.info("Clearing / recreating merge directory: %s" % keyspace_path)
 
-        if os.path.exists(keyspace_path) and os.path.isdir(keyspace_path):
-            logging.warning("Deleteing directory (%s)..." % keyspace_path)
+        if self.overwrite_local_merge and os.path.exists(keyspace_path) and os.path.isdir(keyspace_path):
+            logging.warning("======> DELETING AND OVERWRITING EXISTING MERGE PATH (%s)..." % keyspace_path)
             shutil.rmtree(keyspace_path)
 
         for table in tables:
@@ -231,7 +231,7 @@ class RestoreWorker(object):
 
         thread_pool = Pool(pool_size)
 
-        if self.local_source and not self.overwrite_local_source:
+        if self.local_source:
             meth = self._copy_key
 
         else:
@@ -274,8 +274,6 @@ class RestoreWorker(object):
 
     def _download_key(self, key):
         path = key.name
-        if self.local_source:
-            path = os.path.join(self.local_source, key.name)
         dst = self.dst_from_key(path=path)
         logging.info("downloading %(key)s to %(filename)s" % dict(key=key.name, filename=dst))
         key.get_contents_to_filename(dst)
